moved IntervalsLegendColorMapRow.svelte
