+++ src/components/controls/IntervalsLegendColorMapRow.svelte
